NINJA_STATUS_SLEEP: Avoid too frequent status updates
A build log on a dumb terminal can be big, even for a fast build. Reduce
the output on dumb terminal by printing less frequent if
NINJA_STATUS_SLEEP is set.

Use NINJA_STATUS_SLEEP=inf for total silence.

This is especially useful for build automation servers where the
terminal usually is dumb. You will be able to see progress, but do not
have to overflow the console output with status lines.