Centralize getopt reset
getopt reset is done sometimes with "optind = 0", sometimes with
"optind = 1".  Unfortunately, neither is portable.  For example,
if the bundled getopt see "optind = 1" on the very first call it will
implicitly add a dash in front of the first argument; and while some
glibc versions even crashed if reset with "optind = 1", Apple's
getopt did the same for "optind = 0".

Since we know that there will always be two getopt calls only, of
which the first will be in ReadFlags, place the ugly code there
so that it is not replicated all over ninja.cc.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>