migrate Plan::ready_ to a custom optionally sorted struct
Modifying a key in associative containers leads to UB. This is an
std::set reimplementation with worse characteristics (such as re-sorting
the container on every push), but one that allows for elements to be
out-of-order (which can happen after a key was modified)

This is a bad solution to a problem that ideally should be tackled by
making sure that once Plan is assembled, the graph is never modified.
But this would need massive constifications and rework in the codebase
that I am not familiar with, so let's solve it the dirty way.

Also adds tests for the class.

Signed-off-by: Konstantin Kharlamov <Hi-Angel@yandex.ru>