Support for the question option
This change implements the -q (question) option which is analogous to
Make: when set Ninja doesn't produce any output and exit with zero
status if there's no work to do, non-zero otherwise. This is useful in
scenarios such as continuous integration build as a sanity check to
ensure there are no dirty targets after finishing the build (which will
happen if some outputs/dependencies aren't correctly tracked).