ManifestParser: Avoid calling CanonicalizePath for known nodes
We have been calculating the canonical paths of nodes even if we have
seen them already. As CanonicalizePath is a hot spot it was visible in
the performance measurements. Avoid that by calculating the canonical
path only if the node is not known already.

On a 420MB build.ninja file reduces the noop load time from 5.5s to 5.0s

Signed-off-by: Dimitar Yordanov <dimitar.yordanov@sap.com>