Pull Spellcheck from State and put it into tools
This begins the task of separating out functions that
are current on state that should not be when the implementation
is hidden. This spellcheck function is solidly in the
realm of "things that belong to the user interface". It's also
short and depends on very little, so a good candidate for an
initial move to establish the pattern.