Move deps and build log to tools file.
This pulls out the functions responsible for loading the deps log
and the build log. The deps log and build log now belong to the
State since it will be the owner of most of the state that should
be considered internal to ninja. This patch, as-is, is fairly
incomplete as the goal was to move ownership and the functions
around as simply as possible. The cost is that I've dirtied
some public headers with non-public headers for the time being.
Successive commits will work to clean this up.