Move DiskInterface and BuidLogUser to State.
The main goal here is to remove build_log_user.h from the headers
included from public/tools.h. In order to do that I had to move
BuildLogUser::IsPathDead to state. This makes sense because it is
effectively a question for the current state. However, to keep
the log messages working and to avoid writing directly to stderr
inside ninja I created a new interface for defining a way to handle
log messages from within State. This allows for programs that
use libninja to define their own way to display errors and warnings.