handle EAGAIN when spawning new processes
In Linux it is possible for various syscalls to return EAGAIN when the
system is too busy to handle more load. This is sad. Currently ninja
will just stop in the event of an EAGAIN error. This change allows ninja
to backoff a bit and continue trying to get work done. Note that we have
seen this issue crop up when various parallel build systems are running
concurrently on a single machine.