Split internal and external state, introduce Execution.
The purpose of the Execution is to provide the main point
of entry and control for the library portion of ninja. It
will be where the client of the library will set various
options like the tool or the behavior of the tool. This
means it needs to take back some responsibilities that I
had recently put on state.

This also means that any public headers need to stop knowing
about the internal state as anything more than a pointer.
The disk interface is hard to fully extract, so for now I've
added an escape hatch so that clients of Execution can
get access to the disk interface directly.

I've also removed the ninja command from the state to resolve
a potential circular dependency. It's now kept on the
execution and is required to be set in order for browse
to work correctly.