Make Execution::Options.
The idea is that we have a single class where we are gathering all
of the options that will be used in whatever tool gets executed.
This is to avoid the previous situation where the UI layer is creating
two different objects at the same time - the BuildConfig and
Options - that have essentially the same meaning from the library
client's point of view.

In doing this I reworked the function to parse the flags so that
they just manipulate the Options that will be passed to the Execution.
This means that some error handling functions will need to move
out of the flag parser eventually. I've already done that a bit
for the tool selector, but more will need to be done.