Make Execution::state_ private/protected.
This change got complicated. It started out as a desire to make
state protected within Execution and to provide a read-only accessor
to other functions. Unfortunately the tools module was relying
rather heavily on the fact that it could modify the state by
reaching into the execution, as was the ui::Execute function.

To deal with these I moved most of the ui::Execute logic into Execution.
This is a bit premature - I hadn't yet finished factoring out the
output mechanisms from that code. I also moved some of the logic
for different tools directly into Execution (Browse, Clean, Graph, Query,
Recompact) before they were really ready.

I moved CollectTargets and CollectTargetsFromArgs into UI since that
is eventually where they should live and it made it possible to do some
of this other refactoring.

The next several commits will be cleaning this up and re-separating
out these layers.