Perform parsing of clean arguments during flag parsing.
This moves the logic for parsing the flags for the clean
tool at the same time that other flags are parsed. We want
to do all parsing up front instead of incrementally so that
users of the library can populate the options without
passing through something like flags, but instead by building
up options structures themselves.

This therefore adds a structure for storing the clean tool's
options.