Parse targets for graph during flag parsing.
Continuing the work on doing parsing of commandline args early during procesing
and using the options structure to store that information for the tool to use.
In doing this I've pulled the targets off of Options::Clean so that it can
become common to all of the tools.

I've created a new function for translating from a vector of target names
to a vector of the nodes they refer to which will be used by many tools
that have the same behavior about specifying a list of targets.