Avoid passing argv all the way through browse.
Turns out we only used argv in order to pass the (optional) starting
target to the browse command. That means that we can instead just leverage
the existing machinery for passing the parsed targets.

Also introduces a helpful error message if the user attempts to specify too
many targets. And removes the unnecessary State passed into the Browse function
- I must have added it at some point and no longer need it.