Remove tools:: namespace.
I've finished moving all of the logic into the execution class
which means that we don't need tools anymore. This has the added
benefit of no longer pushing around the argc/argv variables
to so many functions.

In doing this I ripped apart the tools logic and spread it out
among the ui namespace and Execution class. This was imperfect -
I ended up duplicating the kTools list so that I could put
the tool's 'when' property within the Execution class, which cares
about it, rather than in the ui namespace. Eventually I will
get ride of the concept of 'when' entirely so that individual
functions set themselves up correctly. This may make it possible
to avoid having an Execution be a one-shot kind of thing.

I also made the Status a member of Execution rather than creating
it temporarily. This will allow clients of the class to specify
how they would like to get status updates so that they can integrate
more tightly with the class without having to parse stdout.

The other change I made was to special-case the 'list' tool. It
isn't really a tool on an Execution, it's a UI convenience to see
the list of tools. As such it can't have a reasonable implementation
in Execution. So it's special-cased in the ui code.