Create a 'build' tool that matches the patterns of other tools.
Build has previously been special cased as the tool that you get
when you don't choose another tool. That special logic is
littered throughout the ui and Execution code at this point. By
getting rid of the special nature of the tool it's easier to see
how we can break up the 'Execute' function so that each tool's
function can do the work that it needs. This will serve to make
the library API cleaner and less stateful.