Deconstruct Execution::Run.
This function existed primarily to bundle up common code
that was used between the different ninja tools. This commit
un-bundles the common code so that each tool function calls
what it needs. This removes the need for the concept of
when a tool runs. It also allows me to simplify the ui::Execute
function to a level where clients of the library could reasonably
re-implement it.