Turn g_explaining into a data member.
g_explaining was a global flag which caused various parts of the code
to write directly to stdout with debug information. This is a useful
feature, but will be problematic as a library since we don't want
a library indescriminately dumping data to stdout/stderr. Instead
we want it to go through well-defined logging functions so that the
output can be easily captured by clients of the library.