Remove State::Log.
This function was being used by Execution rather than just using its
own pointer to the logger_ because of my previous code that used
unique_ptr. Now that we're not so constrained we can avoid this
pattern which should force me to make each class and function clearly
depend on a logger pointer, not a state pointer.