Fix error handling for getcwd
Quoting from the Linux man page for errno,

"The value in errno is significant only when the return value of the call
indicated an error (i.e., -1 from most system calls; -1 or NULL from most
library functions); a function that succeeds is allowed to change errno. The
value of errno is never set to zero by any system call or library function."

Successful calls to getcwd are allowed to set errno causing the compilation
database not to be written. Spurious failures of this nature were observed on
AIX.

Adjust the error handling for getcwd so that errno is only checked if the call
returned NULL.