Add stream operators to the Logger class.
After fiddling with the logging logic everywhere I've realized that
my strategy for dealing with output is inconsistent. The original
ninja logic has a couple of different classes of uses:

1. Write a warning or error message via printf with a prefix.
2. Stream output from various operations via puts/printf/others.

I need to replace both of these patterns with the Logger class
in such a way that clients can capture the output and do something
useful with it. Until now I've only captured one. I'd also like
to add a distinction between output from the requested ninja operation
and messaging about what ninja is doing and how it may have failed.

The former will go on stdout, the later on stderr, as is convention.