Use logger for status/line printer writing.
This continues on with the work to redirect all output
through the logger class so that clients of the library
can capture it in their own way and deal with it. This
change uses Logger for status printer and the line printer
which is in charge of doing clever things like writing status
lines in-place on VT100 terminals.

Unfortunately in doing this I removed the logic for Windows
which avoids keeping the cursor from moving around everywhere.
It was clever code, but I don't have a windows machine to
test on and even if I did I'm not sure how I would maintain
code of that nature while redirecting through a class
that by design allows clients to avoid sending text to the
console.

This also removes "using namespace std" from the line printer
code, so some std:: prefixes were added.