Reset state when rebuilding.
I had introduced this bug many commits ago - in the original ninja
code the entire state of the build would get re-created any time
it was found that a modification had been made to the build instructions
as part of the build. Now that the build happens as a method on Execution
the state was not being cleared and the build would fail having
re-found most (all) of the same rules in the buildfile.

This commit fixes that by clearing the state as well as the
various classes the state depends on.

I did my best to ascertain whether or not deleting various heap-allocated
objects could be done safely, but there is still some danger here that I got
things wrong, so careful code review would be in order.