graph: batch stat() system calls using io_uring on Linux
This adds an optional dependency on liburing
(https://github.com/axboe/liburing) to use Linux's new io_uring API
which can batch I/O system calls, reducing overhead.

The new source "uring.cc" implements class BulkStat as a wrapper for
liburing, plugged into DependencyScan::RecomputeDirty() which walks
the graph and batches most of the necessary stat() calls before
RecomputeDirty() needs them.

This requires Linux kernel 5.6 or newer.

This patch is not yet integrated well; for example, it bypasses the
DiskInterface, and to avoid breaking the unit tests, this patch adds a
IsReal() method to DiskInterface so we can disable the io_uring code
in the unit tests.