Generate directory for depfile
Ninja creates the directories for all output files, but not the
directory for a depfile. This works for most commands, as the depfile is
typically located in the same directory as the corresponding object file
or the command might generate the directory for it.  However, if it does
not, the command could fail due to directory not existing. This change
creates the directory for the depfile when creating the directories for
the output files.

Fixes #1848