Do not use /bin/sh for execution of built-in commands.
Portable shell programming is a challenge. /bin/sh defaults to bash
on most Linuxes, but not Debian and its descendants (where it points
to dash), which leads to different behaviour of shell builtins,
e.g. `echo`.

Autoconf's guide to portable shell programming calls this situation
a deadlock.

One way to ensure ninja behaviour is identical on all platforms is
require Bash (/bin/bash) in all cases..
Another is to avoid using the shell interpreter altogether.

This patch implements the second alternative.

Fixes #1871