manifest_parser: Add infrastructure for top-level bindings
Top-level bindings in the build manifest allow generators to give ninja
information that is not tied to a specific rule or build statement.  We
already have a two top-level bindings with dedicated implementations:

* builddir: used by NinjaMain
* ninja_required_version: checked directly by parser

Add infrastructure to allow the manifest parser itself to check more
top-level bindings at the end of parsing.