Introduce the ability to escape arbitrary bytes
Presently invocations of programs with multiple lines in an argument
are not obviously possible. This patch introduces the ability to insert
arbitrary bytes into the command with a new $-escape using the
curly-brace form. This is compatible with existing syntax: the `${}`
escape formerly accepted only the `varname` production within the curly
braces which itself doesn't include backslash, so:

  ${\x0a}

inserts an ASCII newline character into the EvalString.
For simplicity and performance, we only allow exactly two hex-bytes.
Bytes less than 0x10 should use a leading zero.

TODO:
  - work out how to best format the status line/-t commands/-v output
    for a command
  - prepare a patch for the documentation
  - add convenience helpers `${\n}`/`${\t}` etc