Always include depfile mtime in build log mtime if special deps binding isn't specified
Since the edge command might produce the depfile (implicit input) along
with the output(s), it should be taken into account when determining the
mtime to record in the build log for the outputs. If it's not, then the
next run could see a stale mtime since the previous run would have
recorded the most recent mtime excluding the depfile itself.

For special depfiles (msvcc and gcc as of this commit), the depfiles are
themselves parsed when the edge command is finished and the most recent
input time of the parsed implicit deps is directly used instead.