dyndep: reconcile dyndep-specified outputs with depfile-specified inputs
When a path loaded from a depfile does not have a node, we create a new
node with a phony edge producing it.  If we later load a dyndep file
that specifies the same node as an output of a known edge, we previously
failed with a "multiple rules generate ..." error.  Instead, since the
conflicting edge was internally generated, replace the node's input edge
with the now-known real edge that produces it.