CLParser: Don't filter filename lines after seeing /showIncludes
The /showIncludes output always comes after cl.exe echos the filename,
so there is no need to filter out filename lines afterwards.

This makes it less likely that CLParser will "over filter" the compiler
output. For example, using the -H flag with clang-cl may lead to output
lines ending in .cc, which are not supposed to be filtered out.