Avoid reallocs in EvalString::Evaluate
Concatenating strings often requires repeated reallocations as the
string grows, plus the cost of copying. These costs are not huge bu
they are easily avoidable when, as in this case, we usually know th
final size.

This saves about 50 ms on a NOP-build of Chromium (not as much as I
had hoped).