Disable stdout buffering in real_main
Previously stdout buffering was disabled in the LinePrinter constructor.
This worked for a long time but ultimately this side-effect caused a
performance bug (issue #2018) in tools such as -t deps. Moving the
disabling of buffering into real_main and only disabling buffering
when a tool is not used makes the desired semantics clearer and restores
the lost performance.

This fixes issue #2018. It has been tested and a 10x speedup was seen
relative to the tip-of-tree version.