Allow linking Ninja binary with extra library.
This patch adds a NINJA_EXTRA_LINK_LIBRARY option to
`CMakeLists.txt`, and a --link-with option to
`configure.py`.

These allow building a version of the Ninja binary that
is linked against a better allocation library.

For example linking against rpmalloc [1] on Linux yields
a notable speedup (10.4 -> 9.1 seconds for a no-op Fuchsia
build with warm filesystem caches).

Example usage:

  mkdir build-cmake && cd build-cmake && rm -rf *
  export LDFLAGS="-lpthread -ldl"
  cmake .. -DCMAKE_BUILD_TYPE=Release -GNinja ... \
           -DNINJA_EXTRA_LINK_LIBRARY=/path/to/librpmalloc_wrapper.a

Or:

  export LDFLAGS="-lpthread -ldl"
  ./configure.py --bootstrap --link-with=/path/to/librpmalloc_wrapper.a

Where the extra linker flags are required by rpmalloc in this case.

[1] https://github.com/mjansson/rpmalloc

Change-Id: I5a50f818f43516ccd82047ce39b6ada156b5232a