Disable re2c from embedding the version number in the source file
The checked in sources (depfile_parser.cc and lexer.cc) have the re2c version
embedded in it. The output didn't change since re2c 1.3, except for the
different version number.

A reproducible build would use a pinned version of re2c anyway, so there
is no need to hardcode its value in the checked in generated files.