Pool: Use std::priority_queue instead of std::set
The priority queue is faster and uses less memory for this
specific use case.

Note that this handles the very important case of duplicate
edges being pushed to the pool: duplicates are removed as
soon as they are detected in Pool::RetrieveReadyEdges().

With a minimual Fuchsia build configuration, this reduces the
time for `ninja -n --quiet` from 5.35s to 4.93s (!!)