configure.py: Support clang-cl with `--platform msvc`
This change allows building Ninja Windows binaries on Linux
with the `clang-cl` toolchain. More specifically:

- `configure.py --platform msvc` will generate a `ninja.build`
  file that invokes `clang-cl` instead of `cl.exe`if the
  right values are defined in CXX and AR environment variables.

- The `--bootstrap` option is not supported. Instead, a host
  version of Ninja is needed to invoke the generated build plan,
  and will compile `ninja.exe` and `ninja_test.exe`.

Note: the generated file works fine under Wine, the `ninja_test.exe`
program fails because mkdtemp() seems to fail in surprising way
under this environment (this looks like a Wine bug, or an issue
with the way Wine is configured on my machine, nothing to do with
Ninja itself).