Fix debug C++03 builds with recent GCC versions.
Fix a runtime error that occurs when building Ninja in debug mode
with -std=c++03 and a recent GCC release. E.g.:

```sh
CXXFLAGS='-std=c++03' ./configure.py --debug --bootstrap && ./ninja_test
```

Will fail with:

```
bootstrapping ninja...
wrote build.ninja.
bootstrap complete.  rebuilding...
[31/31] LINK ninja
[10/384] BuildLogTest.MultiTargetEdge/usr/include/c++/11/debug/vector:445:
In function:
    std::__debug::vector<_Tp, _Allocator>::const_reference
    std::__debug::vector<_Tp,
    _Allocator>::operator[](std::__debug::vector<_Tp,
    _Allocator>::size_type) const [with _Tp = Node*; _Allocator =
    std::allocator<Node*>; std::__debug::vector<_Tp,
    _Allocator>::const_reference = Node* const&; std::__debug::vector<_Tp,
    _Allocator>::size_type = long unsigned int]

Error: attempt to subscript container with out-of-bounds index 0, but
container only holds 0 elements.

Objects involved in the operation:
    sequence "this" @ 0x0x55968842c0a0 {
      type = std::__debug::vector<Node*, std::allocator<Node*> >;
    }
Aborted
```

The root of the problem is that an expression like `&(some_vector[0])`
triggers runtime checks in the debug version of recent libstdc++ libraries,
which cause the failure above.

This issue was previously already worked-around for C++11 with the use of data(),
but this wasn't fixed for C++03. It is assumed here that the GCC/libstdc++ version
used on CI/CQ for C++03 verification are just older and do not include this runtime
check.

Fixing the issue allows developers to verify that their local modifications do
pass the C++03 build in debug mode (used during debugging) without having to install
an older toolchain or rely on CI/CQ.