Add $| escape to represent a literal newline
Currently, ninja variables cannot contain newlines
because the syntax doesn't have any way to express
them.

It's reasonably common for build systems to want to
write out files during the build. The easiest way
to do that with ninja would be to make a rule that
has an rspfile with the desired content, and then
moves the rspfile to the destination location.
However, currently, the rspfile can't contain
newlines, and thus some form of escaping needs to
be done, and the rspfile will need to be copied,
doing the unescaping during the copy, instead of moved.
This hurts build performance. An alternative
implementation would be to use echo commands to write
the file, but that can also have slight performance
penalties due to needing to split up long command lines
into multiple echo commands, and merge the resulting
files afterwards. Using echo commands also requires
more sophisticated shell escaping than just ninja
escaping.

Add a new escape sequence, $|, that will represent
a literal newline.